assignment 8 stuff
added "faster" versions of remote impls.
added fast servers to run them.
made benchmark test.
