+++ src/main/java/remoteimps/faster/fixes.txt
@@ -1,0 +1,1 @@
+1. changed FileWriter in RemoteDataStore to a BufferedWriter
