+++ src/main/java/server/CoordinatorServer.java
@@ -1,16 +1,13 @@
 package server;
 
 import apis.ComputeEngineAPI;
-import apis.DataStoreAPI;
-import implementations.FileDataStore;
 import implementations.ComputeEngine;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
-import remoteimps.RemoteComputationCoordinator;
-import remoteimps.RemoteDataStore;
+import remoteimps.old.RemoteComputationCoordinator;
+import remoteimps.old.RemoteDataStore;
 
 import java.io.IOException;
-import java.rmi.Remote;
 
 public class CoordinatorServer {
 
@@ -20,14 +17,13 @@ public class CoordinatorServer {
 
         RemoteDataStore dataStore = new RemoteDataStore();
 
-        Server server = ServerBuilder.forPort(9090)
+        Server slowServer = ServerBuilder.forPort(2020)
                 .addService(new RemoteComputationCoordinator(computeEngine, dataStore))
                 .build();
 
-        System.out.println("starting server on port 9090...");
-        server.start();
-        server.awaitTermination();
-
-        System.out.println("server stopped");
+        System.out.println("starting slow server on port 2020...");
+        slowServer.start();
+        slowServer.awaitTermination();
+        System.out.println("slow coordinator server stopped");
     }
 }
