+++ src/main/resources/output.txt
@@ -1,6 +1,3 @@
 5,16,8,4,2,1,
 6,3,10,5,16,8,4,2,1,
 7,22,11,34,17,52,26,13,40,20,10,5,16,8,4,2,1,
-5,16,8,4,2,1,
-6,3,10,5,16,8,4,2,1,
-7,22,11,34,17,52,26,13,40,20,10,5,16,8,4,2,1,
